planning: fix compile warning.

+++ modules/planning/math/smoothing_spline/piecewise_linear_constraint.cc
@@ -36,14 +36,14 @@ Eigen::MatrixXd MergeMaxtrices(const std::vector<Eigen::MatrixXd>& matrices) {
   }
   int32_t d = 0;
   for (const auto& mat : matrices) {
-    d += mat.rows();
+    d += static_cast<int32_t>(mat.rows());
   }
-  int32_t col = matrices.front().cols();
+  int32_t col = static_cast<int32_t>(matrices.front().cols());
   Eigen::MatrixXd res = Eigen::MatrixXd::Zero(d, col);
   int32_t index = 0;
   for (const auto& mat : matrices) {
     res.block(index, 0, mat.rows(), mat.cols()) = mat;
-    index += mat.rows();
+    index += static_cast<int32_t>(mat.rows());
   }
   return res;
 }
